+++ config/splat.pspeu.stno4.yaml
@@ -61,7 +61,7 @@ segments:
       - [0x120E0, c, no4/e_toad]
       - [0x12920, c, no4/e_frog]
       - [0x131F0, c, no4/e_particles]
-      - [0x14780, c, no4_psp/e_collect]
+      - [0x14780, c, no4/e_collect]
       - [0x189D0, c, no4/e_misc]
       - [0x190D8, c, no4/collision]
       - [0x1B2B0, c, no4/water_effects]
@@ -106,7 +106,7 @@ segments:
       - [0x68AF8, .rodata, no4/e_red_door]
       - [0x68B10, .rodata, no4/e_fishhead]
       - [0x68B28, .rodata, no4/e_toad]
-      - [0x68B48, .rodata, no4_psp/e_collect]
+      - [0x68B48, .rodata, no4/e_collect]
       - [0x68BA8, .rodata, no4/e_frozen_shade]
       - [0x68BE0, .rodata, no4/e_skeleton_ape]
       - [0x68C00, .rodata, no4/e_spear_guard]
@@ -117,7 +117,6 @@ segments:
       - [0x68FB8, .bss, no4/st_debug]
       - [0x68FC0, bss, 68FC0]
       - [0x68FD0, .bss, no4/create_entity]
-      - [0x68FF0, bss, 68FF0]
-      - [0x69038, .bss, no4_psp/e_collect] #BlitChar
+      - [0x68FF0, .bss, no4/e_collect] #BlitChar
       - [0x69050, bss, 69050]
   - [0x69600]
