+++ src/st/no4/e_collect.c
@@ -2,4 +2,11 @@
 #include "no4.h"
 
 #define HEART_DROP_CASTLE_FLAG 0x40
+
+#ifdef VERSION_PSP
+#define GOLD_COLLECT_TEXT                                                      \
+    _S("$1"), _S("$25"), _S("$50"), _S("$100"), _S("$250"), _S("$400"),        \
+        _S("$700"), _S("$1000"), _S("$2000"), _S("$5000"),
+#endif
+
 #include "../e_collect.h"
